% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{as.ordinal}
\alias{as.ordinal}
\title{Converts a number to an ordinal (e.g., first, second, etc.)}
\usage{
as.ordinal(num)
}
\arguments{
\item{num}{numeric. The number to convert to ordinal.}
}
\description{
Converts a number to an ordinal (e.g., first, second, etc.)
}

