% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{batch maps a function to a batched version of that function.}
\usage{
batch(batch_fn, splitting_strategy = NULL, combination_strategy, size = 50,
  verbose = TRUE, stop = TRUE)
}
\arguments{
\item{batch_fn}{function. The method to batch over.}

\item{splitting_strategy}{function. The strategy used to split up inputs.}

\item{combination_strategy}{function. The strategy used to recombine batches.}

\item{size}{numeric. The size of the packets. Default 50.}

\item{verbose}{logical. Whether or not to announce progress by printing dots.}

\item{stop}{logical. Whether to stop if an error is raised.}

\item{...}{additional arguments to pass to \code{batch_fn}.}
}
\description{
batch maps a function to a batched version of that function.
}

