% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/batchman-package.r, R/batchman.R
\docType{package}
\name{batchman}
\alias{batchman}
\alias{batchman-package}
\title{batchman.}
\usage{
batchman(batch_fn, inputs, ..., size = 50, verbose = TRUE)
}
\arguments{
\item{batch_fn}{function. The method to batch over.}

\item{...}{additional arguments to pass to \code{batch_fn}.}

\item{size}{numeric. The size of the packets. Default 50.}

\item{verbose}{logical. Whether or not to announce progress by printing dots.}

\item{inputs.}{The inputs to split into batches.}
}
\description{
batchman.

Wrap a method and send to it in batches.
}

