% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_by_list}
\alias{combine_by_list}
\title{Combine multiple objects into one object, regardless of class.}
\usage{
combine_by_list(combination_list)
}
\arguments{
\item{combination_list}{list. A list of batches to combine.}
}
\value{
an object of the same class as original, but now including batch.
}
\description{
Combine multiple objects into one object, regardless of class.
}

