% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_orchestration.R
\name{get_batch_filenames}
\alias{get_batch_filenames}
\title{Get path to all batch files in the batches folder}
\usage{
get_batch_filenames(path = "./batches/")
}
\arguments{
\item{path}{Path to the folder containing batch.yaml files.}
}
\value{
vector containing batch filenames.
}
\description{
Get path to all batch files in the batches folder
}
\examples{
get_batch_filenames()
}
