% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_before_fn}
\alias{get_before_fn}
\title{A helper method to get the pre-batched function of a batched function.}
\usage{
get_before_fn(fn)
}
\arguments{
\item{fn}{function. The batched function to look for.}
}
\description{
A helper method to get the pre-batched function of a batched function.
}

