% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_orchestration.R
\name{initialize_batch}
\alias{initialize_batch}
\title{Initialize Batch; creates or updates registry of processes}
\usage{
initialize_batch(batch_filename, update_mode = "update")
}
\arguments{
\item{batch_filename}{Name of the yaml batch file to be initialized.}

\item{update_mode}{Indicates the behavior to adopt if a batch was already initiated in the past.}
}
\value{
logical indicating if all processes within batch have been executed.
}
\description{
Initialize Batch; creates or updates registry of processes
}
\examples{
initialize_batch("demo.yaml", "update")
}
