% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.batched_fn}
\alias{is.batched_fn}
\title{A helper method to determine if a function is already batched.}
\usage{
is.batched_fn(fn)
}
\arguments{
\item{fn}{The function to test.}
}
\description{
A helper method to determine if a function is already batched.
}

