% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.emptyrun}
\alias{is.emptyrun}
\title{A helper method to determine if an empty batch was scheduled to run.}
\usage{
is.emptyrun(x)
}
\arguments{
\item{batches}{The object to see if it contains batches.}
}
\description{
A helper method to determine if an empty batch was scheduled to run.
}

