% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{list_instrument_filenames}
\alias{list_instrument_filenames}
\title{List filenames containing a specific word.}
\usage{
list_instrument_filenames(
  dataset_name,
  contains_term,
  dataset_path = "../data/"
)
}
\arguments{
\item{dataset_name}{The name of the dataset folder in which to search for files. It is assumed that this folder is inside the `data` folder in the project root folder.}

\item{contains_term}{A string specifying the term the returned list of files must contain.}
}
\value{
vector of filenames.
}
\description{
List filenames containing a specific word.
}
\examples{
list_instrument_filenames()
}
