% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_templates.R
\name{new_batch}
\alias{new_batch}
\title{Create new batch file (boilerplate)}
\usage{
new_batch(batch_name)
}
\arguments{
\item{batch_name}{the name of the process to be created. Determines the name of the yaml file to be created.}
}
\value{
None
}
\description{
Create new batch file (boilerplate)
}
\examples{
new_batch("demo")
}
