% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_templates.R
\name{new_process}
\alias{new_process}
\title{Create new process file (boilerplate)}
\usage{
new_process(process_name = "process_name", path = "../")
}
\arguments{
\item{process_name}{the name of the process to be created. Determines the name of the R file and the name of its functions.}
}
\value{
None
}
\description{
Create new process file (boilerplate)
}
\examples{
new_process("demo")
}
