% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{print.batched_function}
\alias{print.batched_function}
\title{Print batched functions as they once were.}
\usage{
\method{print}{batched_function}(fn)
}
\arguments{
\item{fn}{function. The function to print.}
}
\description{
Print batched functions as they once were.
}

