% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_orchestration.R
\name{process_wrapper}
\alias{process_wrapper}
\title{Wrapper function to track and monitor the execution of a process}
\usage{
process_wrapper(filename, process_name, output_dir)
}
\arguments{
\item{filename}{path to file to be processed.}

\item{process_name}{name of the function that does the processing.}

\item{output_dir}{path to file that will store the results.}
}
\value{
NA
}
\description{
Wrapper function to track and monitor the execution of a process
}
\examples{
process_wrapper(filename='./subject_01/response.csv', process_name='process_q1', output_dir = './subject_01/results/')
}
