% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Stores partial progress.}
\usage{
progress()
}
\description{
If a batch breaks, you don't want to lose all your progress done up until
that batch.  Batchman allows partial progress to be stored and retrieved.
}

