% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/recombine.R
\name{recombine}
\alias{recombine}
\title{Recombine a batch onto the original, for all classes.}
\usage{
recombine(original, batch)
}
\arguments{
\item{original.}{The object to get merged upon by the batch.}

\item{batch.}{The batch to merge onto the original. Must be the same class as
original.}
}
\value{
an object of the same class as original, but now including batch.
}
\description{
Recombine a batch onto the original, for all classes.
}

