% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_orchestration.R
\name{run_batch}
\alias{run_batch}
\title{Run a batch of processes}
\usage{
run_batch(batch_filename, batch_size = 100)
}
\arguments{
\item{batch_filename}{Path to the batch of processes to be executed.}

\item{batch_size}{Number of files to be processed per call (default=100).}
}
\value{
None
}
\description{
Run a batch of processes
}
\examples{
run_batch("demo.yaml", 1000)
}
