% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_session_info}
\alias{save_session_info}
\title{Save information about platform and packages used during session}
\usage{
save_session_info(output_dir = "./")
}
\arguments{
\item{output_dir}{path to directory in which to save the session info.}
}
\value{
hash/unique id of the session
}
\description{
Save information about platform and packages used during session
}
\examples{
save_session_info("./")
}
