% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_orchestration.R
\name{select_process}
\alias{select_process}
\title{Select files to processes next from the process registry file.}
\usage{
select_process(registry_filename, batch_size = 3)
}
\arguments{
\item{registry_filename}{path to the process registry file.}

\item{batch_size}{number of files to be processed at once when calling the execution function.}
}
\value{
vector of filenames
}
\description{
Select files to processes next from the process registry file.
}
\examples{
select_process(registry_filename, batch_size = 3)
}
