% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_orchestration.R
\name{start_batchman}
\alias{start_batchman}
\title{Create folders to contain batch and process files}
\usage{
start_batchman(output_dir = "../")
}
\arguments{
\item{output_dir}{Path to output/project folder}
}
\value{
None
}
\description{
Create folders to contain batch and process files
}
\examples{
start_batchman(output_dir = '../')
}
