% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_orchestration.R
\name{update_registry}
\alias{update_registry}
\title{Updates the process registry file}
\usage{
update_registry(registry_filename, source_filenames, keys, values)
}
\arguments{
\item{registry_filename}{path to the process registry file.}

\item{source_filenames}{vector of the filenames to be processed.}

\item{keys}{keys of the process registry table to be updated.}

\item{values}{corresponding values to be used for the update.}
}
\value{
None
}
\description{
Updates the process registry file
}
\examples{
update_registry(..., keys = 'status', values = 'done')
}
