% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{batch maps a function to a batched version of that function.}
\usage{
batch(batch_fn, inputs, ..., splitting_strategy, combination_strategy,
  size = 50, verbose = TRUE, stop = FALSE)
}
\arguments{
\item{batch_fn}{function. The method to batch over.}

\item{...}{additional arguments to pass to \code{batch_fn}.}

\item{size}{numeric. The size of the packets. Default 50.}

\item{verbose}{logical. Whether or not to announce progress by printing dots.}

\item{stop}{logical. Whether to stop if an error is raised.}

\item{inputs.}{The inputs to split into batches.}
}
\description{
batch maps a function to a batched version of that function.
}

