% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{batch maps a function to a batched version of that function.}
\usage{
batch(batch_fn, keys, splitting_strategy = NULL,
  combination_strategy = batchman::combine, size = 50, verbose = TRUE,
  trycatch = FALSE, stop = FALSE)
}
\arguments{
\item{batch_fn}{function. The method to batch over.}

\item{keys}{vector. The names of the keys within the function to batch.
Can be '...' if one is batching a splat function with no keys.}

\item{splitting_strategy}{function. The strategy used to split up inputs.
Leave NULL to use the versatile default splitting strategy.}

\item{combination_strategy}{function. The strategy used to recombine batches.
Defaults to class-agnostic combination.}

\item{size}{numeric. The size of the packets. Default 50.}

\item{verbose}{logical. Whether or not to announce progress by printing dots.}

\item{trycatch}{logical. Whether to wrap the function in a tryCatch block.
Can be used to store and retrieve partial progress on an error.}

\item{stop}{logical. Whether trycatch should stop if an error is raised.}
}
\description{
batch maps a function to a batched version of that function.
}

