% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine multiple objects into one object, regardless of class.}
\usage{
combine(...)
}
\arguments{
\item{...}{A list of batches to combine.}
}
\value{
an object of the same class as original, but now including batch.
}
\description{
Combine multiple objects into one object, regardless of class.
}

