% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\alias{combine_by_list}
\title{Combine multiple objects into one object, regardless of class.}
\usage{
combine(...)

combine_by_list(combination_list)
}
\arguments{
\item{...}{A list of batches to combine.}
}
\value{
an object of the same class as original, but now including batch.
}
\description{
Combine multiple objects into one object, regardless of class.
}

