% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.batched_fn}
\alias{is.batched_fn}
\title{A helper method to determine if a function is already batched.}
\usage{
is.batched_fn(fn)
}
\arguments{
\item{fn}{function. The function to test.}
}
\description{
A helper method to determine if a function is already batched.
}

