% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.done}
\alias{is.done}
\title{A helper method to determine if batchman is done batching.}
\usage{
is.done(call)
}
\arguments{
\item{call}{The batchman call object.}
}
\description{
A helper method to determine if batchman is done batching.
}

