% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.no_batches}
\alias{is.no_batches}
\title{A helper method to determine if batches doesn't contain batches.}
\usage{
is.no_batches(batches)
}
\arguments{
\item{batches}{The object to see if it contains batches.}
}
\description{
A helper method to determine if batches doesn't contain batches.
}

