% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/robust-batch.R
\name{robust_batch}
\alias{robust_batch}
\title{robust_batch re-runs the batch function to be highly resilient to errors.}
\usage{
robust_batch(batched_fn, ...)
}
\arguments{
\item{batched_fn}{function.  The batched function to re-run more robustly.}
}
\description{
robust_batch re-runs the batch function to be highly resilient to errors.
}

