% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/robust-batch.R
\name{robust_batch}
\alias{robust_batch}
\title{robust_batch re-runs the batch function to be highly resilient to errors.}
\usage{
robust_batch(batched_fn, ..., batchman.verbose = isTRUE(interactive()),
  batchman.retries = 3)
}
\arguments{
\item{batched_fn}{function.  The batched function to re-run more robustly.}

\item{batchman.retries}{integer.  The maximum amount of times to retry processing.}
}
\description{
robust_batch re-runs the batch function to be highly resilient to errors.
}

